Move to trainer all files from training process
